refactor: simplify — extract shared sync/preset logic, fix quality and efficiency issues
Code reuse:
- Extract Plaid sync logic into shared src/lib/sync-plaid.ts (was duplicated
  between cron route and trigger route, ~100 lines each)
- Extract date presets into src/lib/date-presets.ts (was duplicated between
  DateRangeFilter component and transactions page)
- Add formatCategory() utility to src/lib/utils.ts

Quality:
- Add confirmation dialog before account deletion (destructive action)
- Add userId to UPDATE/DELETE WHERE clauses for defense-in-depth
- Wrap account deletion in a DB transaction (transactions + account atomic)
- Use typed source union ("plaid" | "import" | "manual") in settings page
- Replace magic sentinel values (-1, -2) with discriminated union types

Efficiency:
- Combine two sequential queries into one in sync/status (COUNT + MAX)
- Batch-delete removed Plaid transactions with inArray instead of N+1 loop
- Parallelize modified transaction updates with Promise.all
- Update all item cursors with single inArray query instead of per-account loop

Net: -312 lines
