+++ src/app/api/sync/status/route.ts
@@ -2,7 +2,7 @@ import { NextResponse } from "next/server";
 import { auth } from "@/lib/auth";
 import { db } from "@/db";
 import { accounts, transactions } from "@/db/schema";
-import { eq, and, desc, sql } from "drizzle-orm";
+import { eq, and, sql, max } from "drizzle-orm";
 
 export async function GET() {
   const session = await auth();
@@ -30,25 +30,11 @@ export async function GET() {
     });
   }
 
-  // Get the most recent Plaid transaction creation time as proxy for last sync
-  const [latest] = await db
-    .select({
-      createdAt: transactions.createdAt,
-    })
-    .from(transactions)
-    .where(
-      and(
-        eq(transactions.userId, session.user.id),
-        eq(transactions.source, "plaid")
-      )
-    )
-    .orderBy(desc(transactions.createdAt))
-    .limit(1);
-
-  // Count total Plaid transactions
-  const [countResult] = await db
+  // Single query for both count and latest createdAt
+  const [stats] = await db
     .select({
       count: sql<number>`COUNT(*)`,
+      lastSyncedAt: max(transactions.createdAt),
     })
     .from(transactions)
     .where(
@@ -60,7 +46,7 @@ export async function GET() {
 
   return NextResponse.json({
     hasPlaidAccounts: true,
-    lastSyncedAt: latest?.createdAt?.toISOString() || null,
-    transactionCount: countResult?.count || 0,
+    lastSyncedAt: stats?.lastSyncedAt?.toISOString() || null,
+    transactionCount: stats?.count || 0,
   });
 }
